+++ lib/card_connect_gateway.rb
@@ -1,10 +1,10 @@
 require 'card_connect_gateway/version'
 require 'card_connect_gateway/configuration'
-require 'card_connect_gateway/helpers'
 require 'card_connect_gateway/base'
+require 'card_connect_gateway/service'
 require 'card_connect_gateway/base_request'
 require 'card_connect_gateway/base_response'
-require 'card_connect_gateway/base_authorized_request'
+require 'card_connect_gateway/authorized_request'
 require 'card_connect_gateway/authorization'
 require 'card_connect_gateway/capture'
 require 'card_connect_gateway/void'
