--- lib/card_connect_gateway/helpers.rb
@@ -1,18 +1,0 @@
-require 'json'
-
-module CardConnectGateway
-  module Helpers
-    def self.included(base)
-      base.class_eval do
-        def self.parse_response(hash)
-
-          hash.keys.each do |key|
-            hash[(key.to_sym rescue key) || key] = hash.delete(key)
-          end
-          hash
-        end
-
-      end
-    end
-  end
-end
